+++ assets/javascripts/kitten/helpers/utils/parser.js
@@ -5,7 +5,8 @@ export const parseHtml = value => {
   if (!value) return
 
   // We need to escape "<3" common emoji
-  const encodedValue = value.replace('<3', '&lt;3')
+  const encodedValue =
+    typeof value === 'string' ? value.replace('<3', '&lt;3') : value
 
   return new HtmlToReact.Parser().parse(`<span>${encodedValue}</span>`).props
     .children